moving logics to hardware_info
